requirements.txt
langchain
langchain-openai
langchain-pinecone
openai
pinecone-client
streamlit
pypdf
python-dotenv
tiktoken
